Modified Pivot and Reach by making an Actuator class to fix static issue
