+++ src/main/java/frc/robot/subsystems/Actuator.java
@@ -1,0 +1,60 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+package frc.robot.subsystems;
+
+import com.revrobotics.RelativeEncoder;
+import com.revrobotics.spark.SparkBase.PersistMode;
+import com.revrobotics.spark.SparkBase.ResetMode;
+import com.revrobotics.spark.SparkLowLevel;
+import com.revrobotics.spark.SparkMax;
+import com.revrobotics.spark.config.ClosedLoopConfig.FeedbackSensor;
+import com.revrobotics.spark.config.SoftLimitConfig;
+import com.revrobotics.spark.config.SparkMaxConfig;
+
+import edu.wpi.first.wpilibj2.command.SubsystemBase;
+
+public class Actuator extends SubsystemBase {
+    private SparkMax armMotor;
+    private SparkMaxConfig motorConfig; 
+    private static RelativeEncoder armEncoder;
+
+
+    public Actuator(int ID, double P, double I, double D, double MinOutput, double MaxOutput, double FF, double Iz, float UpperSoftLimit){
+
+        armMotor = new SparkMax(ID, SparkLowLevel.MotorType.kBrushless);
+        motorConfig = new SparkMaxConfig(); 
+    
+        motorConfig.inverted(false); 
+        armMotor.configure(motorConfig, ResetMode.kResetSafeParameters, PersistMode.kNoPersistParameters);
+        
+        motorConfig.closedLoop 
+        .feedbackSensor(FeedbackSensor.kPrimaryEncoder) 
+        .p(P) 
+        .i(I) 
+        .d(D) 
+        .outputRange(MinOutput,MaxOutput) 
+        .velocityFF(FF) 
+        .iZone(Iz); 
+
+        armEncoder = armMotor.getEncoder();
+        armEncoder.setPosition(0);
+
+        SoftLimitConfig softLimitConfig = new SoftLimitConfig();
+        softLimitConfig.forwardSoftLimitEnabled(true);
+        softLimitConfig.forwardSoftLimit(UpperSoftLimit);
+        softLimitConfig.reverseSoftLimitEnabled(true);
+        softLimitConfig.reverseSoftLimit(UpperSoftLimit);
+
+        motorConfig.apply(softLimitConfig);
+        armMotor.configure(motorConfig, ResetMode.kResetSafeParameters, PersistMode.kNoPersistParameters);
+    }
+
+    public double getPosition() {
+        return armEncoder.getPosition();
+    }
+    public void moveToPositionWithPID(double position) {
+      armMotor.getClosedLoopController().setReference(position, SparkMax.ControlType.kPosition);
+    }
+}
